replace old reports, remove deprecated components
