+++ html/pfappserver/root/src/views/Reports/_api/index.js
@@ -1,27 +1,40 @@
 import apiCall from '@/utils/api'
 
 export default {
-  reports: params => {
-    if (params.sort) {
+  list: params => {
+    if (params.sort)
       params.sort = params.sort.join(',')
-    } else {
+    else
       params.sort = 'id'
-    }
-    if (params.fields) {
+    if (params.fields && params.fields.constructor === Array)
       params.fields = params.fields.join(',')
-    }
     return apiCall.get('dynamic_reports', { params }).then(response => {
       return response.data
     })
   },
-  report: id => {
+  listOptions: () => {
+    return apiCall.options('dynamic_reports').then(response => {
+      return response.data
+    })
+  },
+  item: id => {
     return apiCall.get(['dynamic_report', id]).then(response => {
       return response.data.item
     })
   },
-  searchReport: body => {
-    return apiCall.post(['dynamic_report', body.id, 'search'], body).then(response => {
-      return response.data.items
+  itemOptions: id => {
+    return apiCall.options(['dynamic_report', id]).then(response => {
+      return response.data
     })
+  },
+  search: body => {
+    if (body.id) {
+      return apiCall.post(['dynamic_report', body.id, 'search'], body).then(response => {
+        return response.data
+      })
+    }
+    else {
+      return new Promise(r => r({ items: [] }))
+    }
   }
 }
