+++ html/pfappserver/root/src/views/Reports/_router/index.js
@@ -1,48 +1,34 @@
 import acl from '@/utils/acl'
 import store from '@/store'
-import ReportsIndex from '../'
-import ReportsStore from '../_store'
+import TheIndex from '../'
+import StoreModule from '../_store'
 
-const StandardReportChart = () => import(/* webpackChunkName: "Reports" */ '../_components/StandardReportChart')
-const DynamicReportChart = () => import(/* webpackChunkName: "Reports" */ '../_components/DynamicReportChart')
+const TheView = () => import(/* webpackChunkName: "Reports" */ '../_components/TheView')
 
 const route = {
   path: '/reports',
   name: 'reports',
-  redirect: '/reports/standard/chart/os',
-  component: ReportsIndex,
+  redirect: `/reports/${encodeURIComponent('Accounting::Node')}`,
+  component: TheIndex,
   meta: {
     can: () => acl.$some('read', ['reports']), // has ACL for 1+ children
     isFailRoute: true,
     transitionDelay: 300 * 2 // See _transitions.scss => $slide-bottom-duration
   },
   beforeEnter: (to, from, next) => {
+console.log('before')
     if (!store.state.$_reports) {
       // Register store module only once
-      store.registerModule('$_reports', ReportsStore)
+      store.registerModule('$_reports', StoreModule)
     }
     next()
   },
   children: [
     {
-      path: 'standard/chart/:path([a-zA-Z0-9-_/]+)/',
-      name: 'standardReportChart',
-      component: StandardReportChart,
-      props: (route) => ({ path: route.params.path }),
-      meta: {
-        can: 'read reports'
-      }
-    },
-    {
-      path: 'dynamic/chart/:id([a-zA-Z0-9-_ ]+)',
-      name: 'dynamicReportChart',
-      component: DynamicReportChart,
-      props: (route) => ({ storeName: '$_reports', id: route.params.id, query: route.query.query }),
-      beforeEnter: (to, from, next) => {
-        store.dispatch('$_reports/getReport', to.params.id).then(() => {
-          next()
-        })
-      },
+      path: ':id([a-zA-Z0-9]+.+[a-zA-Z0-9]+)/',
+      name: 'report',
+      component: TheView,
+      props: route => ({ id: route.params.id }),
       meta: {
         can: 'read reports'
       }
